+++ src/libduc/duc.h
@@ -155,7 +155,7 @@ char *duc_dir_get_path(duc_dir *dir);
 void duc_dir_get_size(duc_dir *dir, struct duc_size *size);
 size_t duc_dir_get_count(duc_dir *dir);
 struct duc_dirent *duc_dir_find_child(duc_dir *dir, const char *name);
-int duc_dir_seek(duc_dir *dir, off_t offset);
+int duc_dir_seek(duc_dir *dir, size_t offset);
 int duc_dir_rewind(duc_dir *dir);
 int duc_dir_close(duc_dir *dir);
